move helper functions to the commen/helper.py
